+++ routes/tasks/delete.php
@@ -1,19 +1,21 @@
 <?php
 
+$current_user = get_current_user_from_jwt();
+if ($current_user) {
 
-$id = $_GET['id'];
 
-$deleted = delete_task($id);
+    $id = $_GET['id'];
 
-if ($deleted) {
-     // success but not returning any content
-    http_response_code(204);
+    $deleted = delete_task($id);
 
+    if ($deleted) {
+        // success but not returning any content
+        http_response_code(204);
+    } else {
+        http_response_code(404);
+        echo json_encode('error');
+    }
 } else {
     http_response_code(404);
-    echo json_encode('error'); 
+    echo json_encode('error');
 }
-
-
-
-?>
